handling case where app is stopped when it isnt running

+++ 01-Login/scripts/stop.js
@@ -2,9 +2,13 @@ const ps = require('ps-node');
 ps.lookup({
     command: /node/,
     arguments: /react-scripts\/scripts\/start.js/
-}, (err, result) => {
+}, (err, result = []) => {
     if (err) {
         throw err;
+    }
+    if (!result.length) {
+      process.exit();
+      return;
     }
     const { pid } = result[0];
     ps.kill(pid, (err) => {
